추가 과제 머지 중 추가된 js 파일들 삭제
