+++ packages/assignment-2/src/advanced/advanced.jsx
@@ -1,6 +1,16 @@
 import { createContext, useContext, useState } from "react";
 
-export const memo1 = (fn) => fn();
+const cache = new Map();
+
+export const memo1 = (fn) => {
+  if (cache.has(fn)) {
+    return cache.get(fn)
+  };
+  const ret = fn()
+  cache.set(fn, ret);
+
+  return ret;
+};
 
 export const memo2 = (fn) => fn();
